Speed improvements, particularly when using note-based styles
This update to the protocol used to communicate between LibreOffice and Zotero uses numeric indices to identify citations, rather than the raw field code. This decreases both the amount of data sent via the socket and the amount of time it takes to process it. Additionally, we no longer send calls to getNoteIndex() over the link for each footnote citation. This should provide a significant speed boost for note-based styles.